:truck: rename: 변수명 수정
