+++ src/main/java/com/server/tori/controller/LandmarkLikeController.java
@@ -3,7 +3,7 @@ package com.server.tori.controller;
 import com.server.tori.dto.Landmark.LandmarkLikeResponseDto;
 import com.server.tori.dto.Landmark.LandmarkLikeRequestDto;
 import com.server.tori.dto.Landmark.MyLandmarkResponseDto;
-import com.server.tori.dto.LandmarkDeleteRequestDto;
+import com.server.tori.dto.Landmark.LandmarkDeleteRequestDto;
 import com.server.tori.service.LandmarkLikeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
