+++ src/main/java/com/server/tori/dto/Landmark/LandmarkDetailGetResponseDto.java
@@ -4,22 +4,31 @@ import com.server.tori.entity.Landmark.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-import javax.swing.tree.TreeNode;
 import java.util.List;
 import java.util.stream.Collectors;
 
 @AllArgsConstructor
 @Getter
-public class LandmarkViewResponseDto {
-    private Long landmarkId;
+public class LandmarkDetailGetResponseDto {
     private String name;
+    private String address;
+    private String description;
+    private String price;
+    private String time;
+    private String site;
+    private String phoneNum;
     private List<String> categoryList;
     private List<String> locationList;
     private List<String> imageList;
 
-    public LandmarkViewResponseDto(Landmark landmark, Translation translation) {
-        this.landmarkId = landmark.getId();
+    public LandmarkDetailGetResponseDto(Landmark landmark, Translation translation) {
         this.name = translation.getName();
+        this.address = translation.getAddress();
+        this.description = translation.getDescription();
+        this.price = landmark.getPrice();
+        this.time = landmark.getTime();
+        this.site = landmark.getSite();
+        this.phoneNum = landmark.getPhoneNum();
         this.categoryList = mapCategory(landmark.getCategoryList());
         this.locationList = mapLocation(landmark.getLocationList());
         this.imageList = mapImage(landmark.getImageList());
